requirements.txt
streamlit==1.28.0
openai==0.28.0
python-dotenv==1.0.0
Pillow==10.0.1
pytesseract==0.3.10
pdfkit==1.0.0
streamlit-chat==0.1.0
python-mpesa==0.4.1